Use uint64_t in nvm_ns_info

+++ include/nvm_types.h
@@ -151,9 +151,9 @@ struct nvm_ctrl_info
 struct nvm_ns_info
 {
     uint32_t                ns_id;          // Namespace identifier
-    size_t                  size;           // Size in logical blocks (NSZE)
-    size_t                  capacity;       // Capacity in logical blocks (NCAP)
-    size_t                  utilization;    // Utilization in logical blocks (NUSE)
+    uint64_t                size;           // Size in logical blocks (NSZE)
+    uint64_t                capacity;       // Capacity in logical blocks (NCAP)
+    uint64_t                utilization;    // Utilization in logical blocks (NUSE)
     size_t                  lba_data_size;  // Logical block size (LBADS)
     size_t                  metadata_size;  // Metadata size (MS)
 };
